feat(offchain): enhance `Arena` module
