+++ offchain/Dockerfile.compute
@@ -21,7 +21,7 @@ WORKDIR /app/offchain
 RUN cargo chef cook --release --recipe-path recipe.json
 
 # Build application
-COPY --from=ethereum/solc:stable /usr/bin/solc /usr/bin/solc
+COPY --from=ethereum/solc:0.8.23 /usr/bin/solc /usr/bin/solc
 RUN chmod u+x /usr/bin/solc
 
 COPY ./offchain /app/offchain
