+++ offchain/core/Cargo.toml
@@ -13,8 +13,6 @@ log = { workspace = true }
 base64 = "0.21.3"
 hex = "0.4.3"
 sha3 = "0.10.8"
-serde_json = "1.0.107"
-serde = "1"
 thiserror = "1.0.50"
 
 [build-dependencies]
