+++ offchain/core/src/arena/mod.rs
@@ -1,15 +1,9 @@
-//! This module defines the trait [Arena] that is responsible for the creation and management of
-//! tournaments and one implementation of this trait using the Ethereum blockchain [EthersArena].
+//! This module defines the struct [Reader] that is responsible for the reading the states
+//! of tournaments; and the struct [Sender] that is responsible for the sending transactions
+//! to tournaments
 
 mod config;
 pub use config::*;
 
 mod arena;
 pub use arena::*;
-
-mod ethers_arena;
-pub use ethers_arena::*;
-
-#[doc(hidden)]
-mod util;
-use util::*;
