+++ offchain/core/src/strategy/mod.rs
@@ -1,2 +1,5 @@
+//! This module defines the struct [Player] that is responsible for reacting to the states
+//! of tournaments; and the struct [GarbageCollector] that is responsible for collecting finished matches
+
 pub mod gc;
 pub mod player;
