+++ offchain/dave-compute/src/main.rs
@@ -1,4 +1,4 @@
-use cartesi_compute_core::arena::{ArenaConfig, ContractArtifactsConfig, EthersArena};
+use cartesi_compute_core::arena::{Arena, ArenaConfig};
 use cartesi_compute_core::machine::{CachingMachineCommitmentBuilder, MachineFactory};
 use cartesi_compute_core::strategy::{gc::GarbageCollector, player::Player};
 use ethers::types::Address;
@@ -11,32 +11,18 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     env_logger::init();
     info!("Hello from Dave!");
     let web3_rpc_url = String::from("http://127.0.0.1:8545");
-    let artifacts_path = "/root/offchain/core/artifacts";
     let web3_chain_id = 31337;
-    let contract_artifacts = ContractArtifactsConfig {
-        single_level_factory: format!("{}/SingleLevelTournamentFactory.json", artifacts_path),
-        top_factory: format!("{}/TopTournamentFactory.json", artifacts_path),
-        middle_factory: format!("{}/MiddleTournamentFactory.json", artifacts_path),
-        bottom_factory: format!("{}/BottomTournamentFactory.json", artifacts_path),
-        tournament_factory: format!("{}/TournamentFactory.json", artifacts_path),
-    };
 
-    let player1_private_key =
+    let private_key =
         String::from("0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80");
 
-    let player1_arena_config = ArenaConfig {
+    let arena_config = ArenaConfig {
         web3_rpc_url,
         web3_chain_id,
-        web3_private_key: player1_private_key,
-        contract_artifacts,
+        web3_private_key: private_key,
     };
 
-    let player1_arena = EthersArena::new(
-        player1_arena_config,
-        Some(Address::from_str(
-            "0xdc64a140aa3e981100a9beca4e685f962f0cf6c9",
-        )?),
-    )?;
+    let arena = Arena::new(arena_config)?;
 
     let simple_linux_program =
         String::from("/root/permissionless-arbitration/lua_node/program/simple-program");
@@ -44,29 +30,31 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let machine_rpc_host = "http://127.0.0.1";
     let machine_rpc_port = 5002;
 
-    let player1_machine_factory =
+    let machine_factory =
         MachineFactory::new(String::from(machine_rpc_host), machine_rpc_port).await?;
     let root_tournament = Address::from_str("0xcafac3dd18ac6c6e92c921884f9e4176737c052c")?;
 
-    let mut player1 = Player::new(
-        player1_arena.clone(),
-        player1_machine_factory.clone(),
+    let mut player = Player::new(
+        arena.clone(),
+        machine_factory.clone(),
         simple_linux_program.clone(),
-        CachingMachineCommitmentBuilder::new(player1_machine_factory, simple_linux_program),
+        CachingMachineCommitmentBuilder::new(machine_factory, simple_linux_program),
         root_tournament.clone(),
     );
 
-    let mut player1_gc = GarbageCollector::new(player1_arena, root_tournament);
+    let mut gc = GarbageCollector::new(arena.clone(), root_tournament.clone());
 
     loop {
-        let res = player1.react().await?;
+        let tournament_states = arena.fetch_from_root(root_tournament).await?;
+        let res = player.react(tournament_states).await?;
         if let Some(r) = res {
             info!("Tournament finished, {:?}", r);
             break;
         }
         tokio::time::sleep(Duration::from_secs(1)).await;
 
-        player1_gc.react().await?;
+        let tournament_states = arena.fetch_from_root(root_tournament).await?;
+        gc.react(tournament_states).await?;
         tokio::time::sleep(Duration::from_secs(1)).await;
     }
